% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseqMotifMatch.R
\docType{methods}
\name{calculateNarrowPeaks,ChIPseqMotifMatch-method}
\alias{calculateNarrowPeaks,ChIPseqMotifMatch-method}
\alias{calculateNarrowPeaks}
\title{run MACS2 on the current regions of interest in the bam file using the narrowPeaks options}
\usage{
\S4method{calculateNarrowPeaks}{ChIPseqMotifMatch}(obj)
}
\arguments{
\item{obj}{An object of class ChIPseqMotifMatch}
}
\value{
a (possibly empty) data.frame
}
\description{
run MACS2 on the current regions of interest in the bam file using the narrowPeaks options
}
