% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseqMotifMatch.R
\docType{class}
\name{ChIPseqMotifMatch-class}
\alias{ChIPseqMotifMatch-class}
\alias{.ChIPseqMotifMatch}
\alias{ChIPseqMotifMatch}
\title{ChIPseqMotifMatch-class}
\usage{
ChIPseqMotifMatch(bamFilename, quiet = TRUE)
}
\value{
An object of the ChIPseqMotifMatch class
}
\description{
Survey some or all of a bam file, scoring pwm motif match to sequence
  where ChIP hits have been called by MACS2
}
