% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseqMotifMatch.R
\docType{methods}
\name{getBamFilename,ChIPseqMotifMatch-method}
\alias{getBamFilename,ChIPseqMotifMatch-method}
\alias{getBamFilename}
\title{Get all the name of the bamfile}
\usage{
\S4method{getBamFilename}{ChIPseqMotifMatch}(obj)
}
\arguments{
\item{obj}{An object of class ChIPseqMotifMatch}
}
\description{
Get all the name of the bamfile
}
