% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseqMotifMatch.R
\docType{methods}
\name{getRegionsOfInterest,ChIPseqMotifMatch-method}
\alias{getRegionsOfInterest,ChIPseqMotifMatch-method}
\alias{getRegionsOfInterest}
\title{get the chromosomal regions of interest}
\usage{
\S4method{getRegionsOfInterest}{ChIPseqMotifMatch}(obj)
}
\arguments{
\item{obj}{An object of class ChIPseqMotifMatch}
}
\value{
a (possibly empty) data.frame
}
\description{
get the chromosomal regions of interest
}
