% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPseqMotifMatch.R
\docType{methods}
\name{setRegionsOfInterest,ChIPseqMotifMatch-method}
\alias{setRegionsOfInterest,ChIPseqMotifMatch-method}
\alias{setRegionsOfInterest}
\title{specify the chromosomal regions you wish to examine and operat upon}
\usage{
\S4method{setRegionsOfInterest}{ChIPseqMotifMatch}(obj, tbl.regions)
}
\arguments{
\item{obj}{An object of class ChIPseqMotifMatch}

\item{tbl.regions}{a data.frame with chrom, start, end columns}
}
\description{
specify the chromosomal regions you wish to examine and operat upon
}
